+++ common/src/main/kotlin/computer/obscure/piku/common/ui/classes/Spacing.kt
@@ -1,10 +1,10 @@
 package computer.obscure.piku.common.ui.classes
 
 class Spacing(
-    var left: Float = 0f,
-    var top: Float = 0f,
-    var right: Float = 0f,
-    var bottom: Float = 0f
+    var left: Double = 0.0,
+    var top: Double = 0.0,
+    var right: Double = 0.0,
+    var bottom: Double = 0.0
 ) {
-    constructor(all: Float) : this(all, all, all, all)
+    constructor(all: Double) : this(all, all, all, all)
 }
