+++ common/src/main/kotlin/computer/obscure/piku/common/ui/components/Group.kt
@@ -8,12 +8,12 @@ data class Group(
 ) : Component() {
     override val compType: CompType = CompType.GROUP
 
-    override fun width(): Float {
+    override fun width(): Double {
         val childWidths = props.components.map { it.width() + it.props.pos.x }
         return (childWidths.maxOrNull() ?: 0).toInt() + props.padding.left + props.padding.right
     }
 
-    override fun height(): Float {
+    override fun height(): Double {
         val childHeights = props.components.map { it.height() + it.props.pos.y }
         return (childHeights.maxOrNull() ?: 0).toInt() + props.padding.top + props.padding.bottom
     }
