+++ common/src/main/kotlin/computer/obscure/piku/common/ui/components/Sprite.kt
@@ -9,15 +9,16 @@ open class Sprite(
     override val compType: CompType
         get() = CompType.SPRITE
 
-    override fun width(): Float {
-        val drawWidth = if (props.size.x > 0) props.size.x else (computedSize?.x ?: 0f)
-        val baseWidth = drawWidth + props.padding.left + props.padding.right
+    override fun width(): Double {
+        val drawWidth = if (props.size.x > 0)
+            props.size.x else (computedSize?.x ?: 0)
+        val baseWidth = drawWidth.toDouble() + props.padding.left + props.padding.right
         return (baseWidth * props.scale.x)
     }
 
-    override fun height(): Float {
+    override fun height(): Double {
         val drawHeight = if (props.size.y > 0) props.size.y else (computedSize?.y ?: 0f)
-        val baseHeight = drawHeight + props.padding.top + props.padding.bottom
+        val baseHeight = drawHeight.toDouble() + props.padding.top + props.padding.bottom
         return (baseHeight * props.scale.y)
     }
 }
